Optimize code

+++ src/component.cpp
@@ -435,7 +435,7 @@ static bool displayCompObj(DROID *psDroid, bool bButton, const glm::mat4& modelM
 
 	if (!bButton && psDroid->shieldPoints > 0 && droidGetMaxShieldPoints(psDroid) > 0)
 	{
-		double factor = static_cast<double>(psDroid->shieldPoints) / droidGetMaxShieldPoints(psDroid);
+		float factor = static_cast<float>(psDroid->shieldPoints) / droidGetMaxShieldPoints(psDroid);
 		iShieldPieData = static_cast<SDWORD>(std::round(255.0f * factor));
 		shieldPieFlag = pie_FORCELIGHT | pie_TRANSLUCENT | pie_SHIELD;
 	}
